backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12

sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
psycopg2-binary==2.9.10

python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.1

redis[asyncio]==5.2.1

pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0

stripe==11.3.0
resend==2.4.0
twilio==9.3.7

apscheduler==3.10.4

openpyxl==3.1.5
python-dateutil==2.9.0.post0
chardet==5.2.0

slowapi==0.1.9
httpx==0.28.1

pytest==8.3.4
pytest-asyncio==0.24.0
faker==33.1.0